backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
python-decouple==3.8
celery==5.3.6
redis==5.0.7
channels==4.0.0
channels-redis==4.2.0
daphne==4.0.0
django-filter>=24.2
coreapi==2.3.3
whitenoise==6.5.0
openpyxl==3.1.2
reportlab==4.0.9
lxml==4.9.3
svglib==1.5.1
